Add new promise impl on cljs (make it work in the same way as clj)
This fixes the long lasting issue that on CLJS/JS platform, the promises
cannot be nested because of the some intrinsecs of the underlying
implementation of js/Promesa. Now it works as how it expect to be
working from the beginning.

If you've been relying on the detailed implementation of js/Promise in CLJS,
it's possible that some parts of your code are broken, because now Promise
works the same way in CLJ and CLJS.

Additionally, it adds synchronous introspection of the promise, so from
now, in CLJS you can deref and query the promise state. Obviously it
performs synchoronous but nonblocking operation becase the underlying
platform (JS VM) does not allows blocking operations.